requirements.txt
annotated-types==0.6.0
anyio==4.1.0
astunparse==1.6.3
absl-py==2.0.0
asyncio==3.4.3
attrs==23.1.0
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.12.1
distlib==0.3.7
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
flask==3.0.0
Flask-Cors==4.0.0
flask-session==0.5.0
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.12.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
mediapipe==0.10.8
ml-dtypes==0.2.0
nlp-id==0.1.15.0
nltk==3.8.1
numpy==1.22
oauthlib==3.2.2
openai==1.3.9
opencv-contrib-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
# pip==21.0.1
platformdirs==4.1.0
pluggy==1.3.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
PyMySQL==1.1.0
pyparsing==3.1.1
pytest==7.3.1
python-dateutil==2.8.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
setuptools==49.2.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sounddevice==0.4.6
tensorboard==2.12
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
#tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.6.1
urllib3==2.1.0
virtualenv==20.25.0
werkzeug==3.0.1
wget==3.2
wheel==0.42.0
wrapt==1.14.1
zipp==3.17.0
# cc
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
Pillow==10.1.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
zipp==3.17.0
